add mixins to dataset_dict.py; do not directly export mixins; remove metrics
