+++ src/pie_datasets/__init__.py
@@ -1,24 +1,15 @@
-from .builder import GeneratorBasedBuilder
-from .common import (
-    EnterDatasetDictMixin,
-    EnterDatasetMixin,
-    ExitDatasetDictMixin,
-    ExitDatasetMixin,
-)
+from .builder import ArrowBasedBuilder, GeneratorBasedBuilder
 from .dataset import Dataset, IterableDataset
 from .dataset_dict import DatasetDict
 from .document_formatter import DocumentFormatter
-from .metric import DocumentMetric
 from .statistic import DocumentStatistic
 
 __all__ = [
     "GeneratorBasedBuilder",
+    "ArrowBasedBuilder",
     "Dataset",
     "IterableDataset",
     "DatasetDict",
     "DocumentFormatter",
-    "EnterDatasetMixin",
-    "ExitDatasetMixin",
-    "EnterDatasetDictMixin",
-    "ExitDatasetDictMixin",
+    "DocumentStatistic",
 ]
