--- src/pie_datasets/metrics/__init__.py
@@ -1,3 +1,0 @@
-from .f1 import F1Metric
-
-__all__ = ["F1Metric"]
